* Finalizacion de opcion de Servicios * Creacion de opcion de facturacion

+++ Capa Objetos/Administracion/Inventario/CO_Compras.cs
@@ -14,11 +14,11 @@ namespace Capa_Objetos.Administracion.Inventario
         public string Serie { get; set; }
         public int Id_Proveedor { get; set; }
         public DateTime Fecha_Compra { get; set; }
-        public float Total { get; set; }
+        public decimal Total { get; set; }
         public int Id_Producto { get; set; }
         public int Cantidad { get; set; }
-        public double Precio { get; set; }
-        public double SubTotal { get; set; }
+        public decimal Precio { get; set; }
+        public decimal SubTotal { get; set; }
 
     }
 }
